Step 2 (internal namespace): move reporting/ → internal/reporting/
Per DIR-INTERNAL-NAMESPACE-punit step 2. Eight main + 7 test source
files under org.javai.punit.reporting (Console / Log / Composite
sinks, PUnitReporter, TransparentStatsRenderer,
InfeasibilityMessageRenderer, format helpers) move under
org.javai.punit.internal.reporting.

The verdict.VerdictSink interface stays public in verdict/. Only
the implementations move.

Touched:
- git mv 15 Java files from reporting/ to internal/reporting/.
- Rewrite package declarations.
- Bulk-rewrite import paths across 13 files (excluding architecture
  tests; their string literals are owned by step 5).

Architecture-test bookkeeping:
- reportingUnderInternal rule's freeze-store: empty — at target
  state.
- publicMustNotImportInternal rule: refrozen.

Tests green; punitexamples composite build green.

Closes ArchUnit rule: reportingUnderInternal.
