--- punit-core/archunit_store/8340eb1d-f56f-49d7-853a-f8267e571e40
@@ -1,9 +1,0 @@
-Class <org.javai.punit.reporting.CompositeVerdictSink> does reside in a package 'org.javai.punit.reporting..' in (CompositeVerdictSink.java:0)
-Class <org.javai.punit.reporting.ConsoleVerdictSink> does reside in a package 'org.javai.punit.reporting..' in (ConsoleVerdictSink.java:0)
-Class <org.javai.punit.reporting.DurationFormat> does reside in a package 'org.javai.punit.reporting..' in (DurationFormat.java:0)
-Class <org.javai.punit.reporting.InfeasibilityMessageRenderer> does reside in a package 'org.javai.punit.reporting..' in (InfeasibilityMessageRenderer.java:0)
-Class <org.javai.punit.reporting.LogVerdictSink> does reside in a package 'org.javai.punit.reporting..' in (LogVerdictSink.java:0)
-Class <org.javai.punit.reporting.PUnitReporter$WarningContent> does reside in a package 'org.javai.punit.reporting..' in (PUnitReporter.java:0)
-Class <org.javai.punit.reporting.PUnitReporter> does reside in a package 'org.javai.punit.reporting..' in (PUnitReporter.java:0)
-Class <org.javai.punit.reporting.RateFormat> does reside in a package 'org.javai.punit.reporting..' in (RateFormat.java:0)
-Class <org.javai.punit.reporting.TransparentStatsRenderer> does reside in a package 'org.javai.punit.reporting..' in (TransparentStatsRenderer.java:0)
