+++ punit-core/src/main/java/org/javai/punit/internal/reporting/CompositeVerdictSink.java
@@ -1,4 +1,4 @@
-package org.javai.punit.reporting;
+package org.javai.punit.internal.reporting;
 
 import java.util.List;
 import java.util.Objects;
