+++ punit-core/src/main/java/org/javai/punit/verdict/VerdictTextRenderer.java
@@ -9,8 +9,8 @@ import java.util.Optional;
 import org.javai.punit.api.TestIntent;
 import org.javai.punit.internal.engine.emit.LatencySection;
 import org.javai.punit.verdict.TerminationReason;
-import org.javai.punit.reporting.PUnitReporter;
-import org.javai.punit.reporting.RateFormat;
+import org.javai.punit.internal.reporting.PUnitReporter;
+import org.javai.punit.internal.reporting.RateFormat;
 import org.javai.punit.statistics.ComplianceEvidenceEvaluator;
 import org.javai.punit.statistics.VerificationFeasibilityEvaluator;
 import org.javai.punit.statistics.transparent.StatisticalVocabulary;
