+++ punit-sentinel/src/test/java/org/javai/punit/sentinel/SentinelConfigurationTest.java
@@ -2,8 +2,8 @@ package org.javai.punit.sentinel;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import org.javai.punit.reporting.CompositeVerdictSink;
-import org.javai.punit.reporting.LogVerdictSink;
+import org.javai.punit.internal.reporting.CompositeVerdictSink;
+import org.javai.punit.internal.reporting.LogVerdictSink;
 import org.javai.punit.verdict.VerdictSink;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
